admin_service/requirements.txt
confluent_kafka
psycopg2-binary
flask
flask-cors
gunicorn
python-dotenv 